Added in gear box. Added new UI for the solar panel and cleaned up the solar panel logic to respect rain and sky visibility. Added probability output as a new JEI ingredient type, so tooltips should be cleaner now. JEI will also render the percentage chance and bonus directly on the slot. Couple other minor tweaks and fixes.

+++ src/main/java/theking530/staticcore/gui/drawables/SpriteDrawable.java
@@ -16,14 +16,13 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 import theking530.staticcore.utilities.Color;
 import theking530.staticcore.utilities.Vector2D;
 
-
 @OnlyIn(Dist.CLIENT)
 public class SpriteDrawable implements IDrawable {
 	private ResourceLocation sprite;
 	private Vector2D size;
 	private Color tint;
 
-	public SpriteDrawable(ResourceLocation sprite, int width, int height) {
+	public SpriteDrawable(ResourceLocation sprite, float width, float height) {
 		this.sprite = sprite;
 		this.size = new Vector2D(width, height);
 		this.tint = Color.WHITE;
